requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
tensorflow==2.13.0
numpy==1.24.3
Pillow==9.5.0
python-multipart==0.0.6
scikit-learn==1.3.0
streamlit==1.24.0
requests==2.31.0
opencv-python==4.8.0.74
kagglehub 